+++ example/lib/main.dart
@@ -14,34 +14,26 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
+  String _version = 'Unknown';
 
   @override
   void initState() {
     super.initState();
     initPlatformState();
-    FlutterCellrebelSdk.init("xpqc5glzli");
+    CellRebelSDK.init("CLIENT_KEY");
   }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    // We also handle the message potentially returning null.
+    String sdkVersion;
     try {
-      platformVersion =
-          await FlutterCellrebelSdk.platformVersion ?? 'Unknown platform version';
+      sdkVersion =
+          await CellRebelSDK.getVersion ?? 'Unknown platform version';
     } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
+      sdkVersion = 'Failed to get platform version.';
     }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
     if (!mounted) return;
 
     setState(() {
-      _platformVersion = platformVersion;
+      _version = sdkVersion;
     });
   }
 
@@ -56,13 +48,14 @@ class _MyAppState extends State<MyApp> {
           child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: [
+            Text('Version: $_version\n'),
             ElevatedButton(
               child: Text('Start tracking'),
-              onPressed: FlutterCellrebelSdk.startTracking,
+              onPressed: CellRebelSDK.startTracking,
             ),
             ElevatedButton(
               child: Text('Stop tracking'),
-              onPressed: FlutterCellrebelSdk.stopTracking,
+              onPressed: CellRebelSDK.stopTracking,
             )
           ],
         ),